iniciado home view em html, inicio de modificações dos arquivos controler, service e model student

+++ src/main/java/br/dcc/ufjf/atividades/repository/RepositorySubjects.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.dcc.ufjf.atividades.model.Subject;
 
 @Repository
-public interface RepositorySubjects extends JpaRepository<Subject,Integer>{
+public interface RepositorySubjects extends JpaRepository<Subject,Long>{
 
 }
